requirements.txt
Django>=4.2.1,<4.3
djangorestframework==3.14.0
python-dateutil==22.0.4